Add a minimalistic health check to the Docker image.
This adds a minimalistic check to the Docker image that verifies that
HAProxy is actually listening on port 2375, marking the image as
unhealthy if the check fails. This, in turn, allows container
orchestration tools to notify an administrator that something is wrong
or even restart the container automatically.

This utilizes the Busybox implementation of `netcat`, which is already
included in the image itself, and thus has functionally zero impact on
image size. The check interval is set for once every 60 seconds, which
is reasonably normal for small services like this.